only run once

+++ prebuild_linux.sh
@@ -1,13 +1,16 @@
 #!/bin/bash
 set -e -x
 
-# Install system packages required by psychtoolbox
-yum install -y libusb1-devel libXi-devel alsa-lib-devel jack-audio-connection-kit-devel
+if [! -d 'portaudio/'] # check if we've run this already
+then
+    # Install system packages required by psychtoolbox
+    yum install -y libusb1-devel libXi-devel alsa-lib-devel jack-audio-connection-kit-devel
 
-# build portaudio manually (couldn't find PaAlsa_SetNumPeriods symbol in the system package)
-git clone https://git.assembla.com/portaudio.git
-cd portaudio
-git checkout pa_stable_v190600_20161030
-./configure && make
-make install
-cd ..
+    # build portaudio manually (couldn't find PaAlsa_SetNumPeriods symbol in the system package)
+    git clone https://git.assembla.com/portaudio.git
+    cd portaudio
+    git checkout pa_stable_v190600_20161030
+    ./configure && make
+    make install
+    cd ..
+fi
